Fix next section logic
The code that checked the color scheme of the next section and
then set the curve mask of the current section to that color
was old and didn't work. This caused the curve mask to be the
wrong color. I updated the code to work with Jekyll's version
of Liquid.