Fix gradient
I created a smoother and more even gradient and a better
animation timing function and applied it to the gradient-box
and fadeimg elements. This makes the fade animation look way
better.

I also created a seperate gradient animation and applied it to
only the Holochain logo in the header. This gradient is less
gradual, so the Holochain logo always has around two colors in
it, just like its static version.